Add files via upload
This file was a script I created at Mapware to help automate the GIS workflow as a precursor to the ML Pipeline. The tool creates two new directories with the current date and time, accepts an input raster ortho, removes the no data value associated with the raster (if any), and finally splits the raster in 1-acre tiles that are saved into the directories initially created.